requirements.txt
fastapi>=0.135.0
uvicorn>=0.42.0
httpx>=0.28.0
websockets>=16.0
python-dotenv>=1.2.0
python-multipart>=0.0.22
websocket-client>=1.9.0
